requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bark==0.1.5
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.1.3
contourpy==1.2.1
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.6.7
decorator==5.1.1
deepdiff==8.0.1
defusedxml==0.7.1
distro==1.9.0
einops==0.8.0
emoji==2.13.2
encodec==0.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.115.2
fastjsonschema==2.16.3
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
funcy==2.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gradio==4.44.1
gradio_client==1.3.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.4
importlib-metadata==6.6.0
importlib_resources==6.4.0
inference-sdk==0.10.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.1.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.5
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.129
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.6.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
olefile==0.47
openai==1.51.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.5.0
plotly==5.24.1
praw==7.7.1
prawcore==2.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.15.1
Pympler==1.1
pyparsing==3.1.2
pypdf==5.0.0
PyPDF2==3.0.1
pyrsistent==0.19.3
pythainlp==5.0.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2024.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
RapidFuzz==3.10.0
regex==2024.5.15
requests==2.28.1
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rsa==4.9
ruff==0.6.9
s3transfer==0.10.3
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.2
sendgrid==6.11.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.35
stack-data==0.6.2
starkbank-ecdsa==2.2.0
starlette==0.39.2
streamlit==1.20.0
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
terminado==0.17.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.18.0
tornado==6.3.1
tqdm==4.66.4
traitlets==5.9.0
transformers==4.41.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.19
unstructured==0.15.13
unstructured-client==0.25.9
update-checker==0.18.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.31.1
validators==0.34.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.7
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
